Trait rustcoalescence_algorithms::Algorithm

pub trait Algorithm<'p, M: MathsCore, O: Scenario<M, Self::Rng>, R: Reporter, P: LocalPartition<'p, R>>: Sized + AlgorithmParamters + AlgorithmDefaults {
    type Rng: RngCore<M>;
    type LineageReference: LineageReference<M, O::Habitat>;
    type LineageStore: LineageStore<M, O::Habitat, Self::LineageReference>;
    fn get_logical_partition(
        args: &Self::Arguments, 
        local_partition: &P
    ) -> Partition;
    fn initialise_and_simulate<I: Iterator<Item = u64>>(
        args: Self::Arguments, 
        rng: Self::Rng, 
        scenario: O, 
        pre_sampler: OriginPreSampler<M, I>, 
        pause_before: Option<NonNegativeF64>, 
        local_partition: &mut P
    ) -> Result<SimulationOutcome<M, Self::Rng>, Self::Error>;
    fn resume_and_simulate<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
        args: Self::Arguments, 
        rng: Self::Rng, 
        scenario: O, 
        pre_sampler: OriginPreSampler<M, I>, 
        lineages: L, 
        resume_after: Option<NonNegativeF64>, 
        pause_before: Option<NonNegativeF64>, 
        local_partition: &mut P
    ) -> Result<SimulationOutcome<M, Self::Rng>, ResumeError<Self::Error>>;
    fn fixup_for_restart<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
        args: Self::Arguments, 
        rng: Self::Rng, 
        scenario: O, 
        pre_sampler: OriginPreSampler<M, I>, 
        lineages: L, 
        restart_at: PositiveF64, 
        fixup_strategy: RestartFixUpStrategy, 
        local_partition: &mut P
    ) -> Result<SimulationOutcome<M, Self::Rng>, ResumeError<Self::Error>>;
}

Associated Types

type Rng: RngCore<M>

type LineageReference: LineageReference<M, O::Habitat>

type LineageStore: LineageStore<M, O::Habitat, Self::LineageReference>

Required methods

fn get_logical_partition(
args: &Self::Arguments,
local_partition: &P
) -> Partition

fn initialise_and_simulate<I: Iterator<Item = u64>>(
args: Self::Arguments,
rng: Self::Rng,
scenario: O,
pre_sampler: OriginPreSampler<M, I>,
pause_before: Option<NonNegativeF64>,
local_partition: &mut P
) -> Result<SimulationOutcome<M, Self::Rng>, Self::Error>

Errors

Returns a Self::Error if initialising the fresh simulation or running the algorithm failed

fn resume_and_simulate<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
args: Self::Arguments,
rng: Self::Rng,
scenario: O,
pre_sampler: OriginPreSampler<M, I>,
lineages: L,
resume_after: Option<NonNegativeF64>,
pause_before: Option<NonNegativeF64>,
local_partition: &mut P
) -> Result<SimulationOutcome<M, Self::Rng>, ResumeError<Self::Error>>

Errors

Returns a ContinueError<Self::Error> if initialising the resuming simulation or running the algorithm failed

fn fixup_for_restart<I: Iterator<Item = u64>, L: ExactSizeIterator<Item = Lineage>>(
args: Self::Arguments,
rng: Self::Rng,
scenario: O,
pre_sampler: OriginPreSampler<M, I>,
lineages: L,
restart_at: PositiveF64,
fixup_strategy: RestartFixUpStrategy,
local_partition: &mut P
) -> Result<SimulationOutcome<M, Self::Rng>, ResumeError<Self::Error>>

Errors

Returns a ContinueError<Self::Error> if fixing up the restarting simulation (incl. running the algorithm) failed